♻️ refactor: 품질 검수 기반 리팩토링 - 캘린더, 통계 컴포넌트에서 각각 수행하던 데이터 요청을 메인페이지에서 한번만 요청하도록 변경
